add handling for inconsistent string/list returns
I've found that the underling ldap library returns an inconsistently-typed result for "memberOf" when I use it agsinst my local Active Directory server.  When there's a single group, a string is returned, and we only return a list when there are either 0 or >1 groups.

I've added a new function call to coerce "memberOf" into a list prior to sorting it.